Get textures loaded/rendering, add zoom functionality to the camera.
This performs incredibly badly because it's uploading ~2000 textures to
the graphics card to render this stuff. We could render the map in far
larger chunks than this (honestly the whole thing could just be
pre-rendered, as it doesn't change, but even if we wanted to allow for
that we could do this better). But, this works to get something on the
screen, so checking it in for now.