Rework tilemap chunking to chunk on unique textures instead of position.
This lets us greatly reduce the number of tilemaps in use (which I think
helps limit overhead there), often to just 1. Potentially there could be
some benefit to trying to balance the tilemaps if there are multiple,
but seems complex so not gonna worry about it for now.

This also fixes an issue where Tiles were not marked as children of the
Tilemap, and thus we were leaking a ton of tiles when you loaded new
maps.