Generate code for flingy data instead of loading the file at runtime.
This is something I'm going to do for the rest of the dat files as well
(and iscript + tilesets as well). I think it makes more sense to generate
code for these things that constitute "game rules". It will make it
possible to simulate games without needing the copyrighted graphical files
(which will make things more useful in e.g. WASM environments, on a server,
etc.). It also makes it easier for us to make this data more directly
usable (adding newtypes, relating units -> sprites, etc.) at
compile-time.

This does remove the ability to use existing mod tools to change these
things, but I don't think that's a huge loss given that most active
modders aren't even working with the SC:R formats to begin with. It will
also be more possible for modders to change our code directly if they
wish.