Add tests for marines, ghosts, and tanks, as well as tests on a map dealing with map edges.
As of this commit the unit bounds have an additional pixel added to
their right/bottom values. This makes them work more intuitively (a
[0, 0, 0, 0] bounds will have size 0 as I expected). Blizzard's code
adds 1 to these values extremely frequently, but there may be cases they
do something with the "raw" value. This is mainly something to be aware
of when implementing compatible algorithms.