Refactor construction creation/placement code to use Events + exclusive systems that run on the FixedUpdate schedule, instead of using custom commands.
This seems preferable to me because the custom commands provided no
guarantee of when the creation/placement would run, which made them
risky to use (you need to consider where the system you are using them
from is running, which might also change in the future). With the
events, we just need to to ensure that any creations meant to run during
the current turn run before these exclusive systems (I need to add a
custom schedule or something to make sure these run last probably).