Add basic handling for drag + click selection.
Currently this prefers, in order:

- Units from the current player
- Buildings from the current player
- Constructs from anyone

Whichever group contains any constructs will be sorted by distance from
the start of the selection (or the single point of the click if not a
drag) and have either 12 or 1 entities picked (whichever is the relevant
limit).

I think I need to adjust the last case to add another set of priorities
(to avoid neutral constructs like mineral fields if a non-neutral
construct is available), but I will do that in a follow-up commit.

Would also be very nice to have tests for this but I have not pursued
that yet :)