Implement proper graphics for selected constructs.
The colors are definitely not correct currently, it seems like the game
just ignores the incoming color of the sprite and applies a target color
without blending, which is something bevy's stock sprite stuff doesn't
seem to support. We'll probably need to write a custom shader in the
future for these (and figure out the actual colors being used).