Generate [a lot] of code for iscript.bin.
I think this structure will likely work for iscript playback but can't
really know until I start implementing it.

IscriptCollection is the collection of scripts assigned to particular
Flingy's. They store an array of IscriptIds (with the end trimmed to
remove None values since their presence adds no value there) which map
positionally to IscriptType. I plan to write some helpers to make
retrieving a script by IscriptType easier in the near future.

IscriptIds are an index into ISCRIPT_ANIMS, which has a dummy empty
script prepended to it to make Option<> optimization possible for
IscriptId. ISCRIPT_ANIMS is a packed array of all the anim blocks,
sorted from most to least referenced. This may be a reasonable way to
sort things or may not be (or may not really matter at all), remains
to be seen. Since we are using indexes into a packed array here instead
of unpacked offsets in a binary file, all the references (both in
commands that use an IscriptLabel as well as IscriptIds) must be
rewritten when generating the code.