Implement enough of the iscript system to give constructs shadows.
Not much else of the Iscript stuff is there yet, and shadows aren't
really visually the same as Blizzard's (but it does seem kind of close,
minus shadow stacking). Further work in that direction will require a
custom shader so I'm not pursuing that for now, we can approach it as
part of the work for team colors.

Image state is stored in one big component because it makes for less
things to pass into the iscript controller for execution, potentially we
could split things up more though. I'm trying to avoid giving a mutable
World to the iscript controller if I can, as it will make it pretty
difficult to use I think, but depending on what is needed for
implementing the other iscript instructions, it may be the only real
path forward. We'll see I guess :)

Iscript execution is currently going to be out of order relative to
Blizzard's, we need to make sure to iterate the entities in an order
consistent with them. This same thing will be required for
constructs/sprites/images/etc. so I'm just gonna deal with that when we
are closer to being able to match replay playback.