Implement LO* file usage for CreateGasOverlays (including a Bevy asset loader for this type).
LO* files are loaded whenever an anim file is loaded that references
them. Currently we only load special overlay LO files, but we need to
implement the other 5 types similarly. When an iscript anim is used that
needs this location info, we add a component to the Image entity that
indicates we need to load it. As soon as it is available in the assets,
the offsets will be applied and the indicator component will be removed.

I'm not sure if this will work for every type of LO usage (there are
iscript operations to specify particular LO usage that might need to be
updated from the offsets again each frame or something?), but it does
work for the gas overlays.