Refactor trigger asset load from LoadingAnim creation to happen as a component hook.
This should be a bit nicer since it will kick off the asset load ASAP
instead of waiting until the SpawnScene step, and also insert the
overlay components immediately (I think this might be able to simplify
some of the overlay code? I need to look at that still).

Potentially we should handle on_replace here as well and do the same
thing, but it seems like a bad idea to reuse entities for different
anims to me (I think? I guess we might need to do this for thing like
sieged tanks, unsure), but if we need it it's probably not that crazy to
add.